Neural Network somewhat working.

+++ backend/experiments.py
@@ -19,22 +19,26 @@ from sklearn.metrics import accuracy_score, f1_score, precision_score
 from sklearn.metrics import ConfusionMatrixDisplay
 from sklearn.metrics import confusion_matrix
 
-# Models
+## Models
 from sklearn.svm import SVC
 from sklearn.neighbors import KNeighborsClassifier
 from werkzeug.utils import secure_filename
 
-from backend.calculator import Calculator, read_data
+from backend.calculator import Calculator
 
 # sys.path.append("backend")
 # sys.path.append("Classes")
 # sys.path.insert(1, '/Classes/')
 from backend import MLA
-from backend import Validation
+from backend import MLA_Validation
+from backend import DLANN_Validation
 from backend.Classes.PreoptimizationFiles import Preoptimization
 from backend.Classes.PreoptimizationFiles import Standardization
 from backend.Classes.NeuralNetworkFiles import NeuralNetwork
 from backend.Classes.NeuralNetworkFiles import Core
+from backend.Classes.NeuralNetworkFiles import Compiler
+
+from keras.utils import to_categorical 
 
 from sklearn.model_selection import KFold
 
@@ -45,8 +49,6 @@ logger = logging.getLogger()
 
 # should be modified to work with a db or class if necessary.
 # Otherwise, the choices being pre-available on page load will slightly speed the process.
-
-
 @bp.before_app_request
 def load_possible_experiments():
     list_Algorithms = MLA.getMLAs()
@@ -57,24 +59,22 @@ def load_possible_experiments():
     list_Layer_Categories = NeuralNetwork.getLayer_Categories()
     list_Core = Core.getCore()
 
-    # ----------------Data Types-------------------------->
+    #----------------Data Types-------------------------->
     g.data_types = ["Tabular", "Image", "Text", "Signal"]
     g.data_types_Name = ["Tabular", "Image", "Text", "Signal"]
     g.data_types_Info = [["Tabular is data that is displayed in columns or tables and does not pertain to the other data types."],
-                         ["Image is data that is usually 3-dimentional corresponding to the three color spectrom, RGB"],
-                         ["Text is data containing textual words or sentences."],
-                         ["Signal is data containing time-series data corresponding to a signal format."]]
+                      ["Image is data that is usually 3-dimentional corresponding to the three color spectrom, RGB"],
+                      ["Text is data containing textual words or sentences."],
+                      ["Signal is data containing time-series data corresponding to a signal format."]]
     g.Data_Types = zip(g.data_types, g.data_types_Name, g.data_types_Info)
 
-    # ----------------new Cycon--------------------------->
-    # First choice is between various categories of ML
+    #----------------new Cycon--------------------------->
+    # First choice is between various categories of ML 
     g.section_Method = ["MLA", "DLANN"]
-    g.section_Method_Display_Name = [
-        "Machine Learning Algorithm (MLA)", "Deep Learning Artifical Neural Networks (DLANN)"]
+    g.section_Method_Display_Name = ["Machine Learning Algorithm (MLA)", "Deep Learning Artifical Neural Networks (DLANN)"]
     g.section_Info = [["Machine learning algorithms are mathematical model mapping methods. They are used to learn patterns embedded in the existing training dataset in order to perform pattern recognition, classification, and prediction.\n\nCurrently, the only algorithms available on cycon is under the classification objective. Other objectives that will be added later include clustering and regression."],
                       ["DLANN info"]]
-    g.Methodologies = zip(
-        g.section_Method, g.section_Method_Display_Name, g.section_Info)
+    g.Methodologies = zip(g.section_Method, g.section_Method_Display_Name, g.section_Info)
 
     # Obtain the selection of Algorithm Names and Definition.
     algorithm_Names = []
@@ -89,7 +89,7 @@ def load_possible_experiments():
     g.section_Info = algorithm_Definition
     g.Algorithms = zip(g.section_algorithm, g.section_Info)
 
-    # ----------------Preoptimiztion-------------------------->
+    #----------------Preoptimiztion-------------------------->
     # create list of options to choose a category of preoptimization options.
     preopt_Category_Names = []
     preopt_Category_Display_Names = []
@@ -101,9 +101,8 @@ def load_possible_experiments():
         preopt_Category_Definition.append(preopt_Category.getDefinition())
     g.section_preopt_Category_Names = preopt_Category_Names
     g.section_preopt_Category_Display_Names = preopt_Category_Display_Names
-    g.section_preopt_Category_Info = preopt_Category_Definition
-    g.preopt_Categories = zip(g.section_preopt_Category_Names,
-                              g.section_preopt_Category_Display_Names, g.section_preopt_Category_Info)
+    g.section_preopt_Category_Info =  preopt_Category_Definition
+    g.preopt_Categories = zip(g.section_preopt_Category_Names, g.section_preopt_Category_Display_Names, g.section_preopt_Category_Info)
 
     # create list of options for the first category of preoprtimization (I.E. Standardization)
     preopt_Names = []
@@ -116,11 +115,10 @@ def load_possible_experiments():
         preopt_Definition.append(preopt.getDefinition())
     g.section_preopt_Names = preopt_Names
     g.section_preopt_Display_Names = preopt_Display_Names
-    g.section_preopt_Info = preopt_Definition
-    g.preopts = zip(g.section_preopt_Names,
-                    g.section_preopt_Display_Names, g.section_preopt_Info)
+    g.section_preopt_Info =  preopt_Definition
+    g.preopts = zip(g.section_preopt_Names, g.section_preopt_Display_Names, g.section_preopt_Info)
 
-    # -------------------Neural Network Layers----------------------->
+    #-------------------Neural Network Layers----------------------->
     # create list of options to choose a category of layer options.
     layer_Category_Names = []
     layer_Category_Display_Names = []
@@ -132,9 +130,8 @@ def load_possible_experiments():
         layer_Category_Definition.append(layer_Category.getDefinition())
     g.layer_Category_Names = layer_Category_Names
     g.layer_Category_Display_Names = layer_Category_Display_Names
-    g.layer_Category_Info = layer_Category_Definition
-    g.NN_layer_Categories = zip(
-        g.layer_Category_Names, g.layer_Category_Display_Names, g.layer_Category_Info)
+    g.layer_Category_Info =  layer_Category_Definition
+    g.NN_layer_Categories = zip(g.layer_Category_Names, g.layer_Category_Display_Names, g.layer_Category_Info)
 
     # create list of options for the first category of preoprtimization (I.E. Core)
     layer_Names = []
@@ -147,8 +144,9 @@ def load_possible_experiments():
         layer_Definition.append(layer.getDefinition())
     g.layer_Names = layer_Names
     g.layer_Display_Names = layer_Display_Names
-    g.layer_Info = layer_Definition
+    g.layer_Info =  layer_Definition
     g.layers = zip(g.layer_Names, g.layer_Display_Names, g.layer_Info)
+
 
 
 # Old LCA for testing purposes, REMOVE BEFORE SUBMITTING.
@@ -157,17 +155,13 @@ def LCA_Old():
     return render_template("experiments/LCA_Old.html")
 
 # REMOVE ASAP
-
-
 @bp.route("/new_experiment")
 def new_experiment():
     return render_template("experiments/new_experiment.html")
 
-
 @bp.route("/lca")
 def lca():
     return render_template("experiments/lca.html")
-
 
 @bp.route("/cycon")
 def cycon():
@@ -180,51 +174,34 @@ def calculate():
     if not processes_input:
         return flash("Invalid arguments!")
     processes = json.loads(processes_input)
-    data = dict()
-    for key, value in request.files.items():
-        logger.debug("=== file data: %s %s", key, value)
-        data[key], error = read_data(value)
-        if error:
-            return jsonify(error=error)
-    logger.debug("=== Equations: %s %s", processes, data)
+    logger.debug("=== Equations: %s %s", processes, processes_input)
+    calculator = Calculator()
 
     results = []
     num_process = 0
     for process in processes:
         num_process += 1
         filename = process.get("filename")
-        if not filename:
-            error = f"Missing filename for process {num_process}"
-            return jsonify(error=error)
-        process_data = data.get(filename)
-        if process_data is None:
-            error = f"Missing csv file '{filename}' for process {num_process}"
-            return jsonify(error=error)
+        csv_file = request.files.get(filename)
         equation = process.get("equation")
         if not equation or equation == "":
             error = f"Missing equation for process for process {num_process}"
             return jsonify(error=error)
         name = process.get("name")
-        logger.debug("=== Process %s: %s %s",
-                     num_process, filename, process_data)
-        calculator = Calculator()
-        evaluated, result = calculator.evaluate(
-            equation=equation, data=process_data)
-        logger.debug("=== Process %s Results: %s, %s",
-                     num_process, evaluated, result)
-        if not evaluated:
-            error = f"Error in the formula for process {num_process}\n{result}"
-            return jsonify(error=error)
-
-        output = dict(result=result, name=name, equation=equation)
+        logger.debug("=== Process (file): %s %s %s", filename, csv_file, request.form)
+        evaluated, result = calculator.evaluate(equation=equation, csv_file=csv_file)
+        output = dict(result=result, name=name)
         results.append(output)
+
+        if not evaluated:
+            flash(str(result))
 
     logger.debug("=== results: %s | dtype %s", results, type(results))
     return jsonify(processes=results)
 
 
 @bp.route("/run_experiment", methods=["POST"])
-def run_experiment():
+def run_experiment():    
     processes_input = request.form.get("processes")
     data = json.loads(processes_input)
 
@@ -234,41 +211,69 @@ def run_experiment():
     data["csvFileName"] = filename
     data["csvFile"] = csv_file
 
-    # Split
-    if data['validation'] == "Split":
-        status, msg, Metrics = Validation.Split(data)
+    # Perform MLA if chosen
+    if data['methodology'] == "MLA":
+        # Split
+        if data['validation'] == "Split":
+            status, msg, Metrics = MLA_Validation.Split(data)
 
-    # K-Fold
-    elif data['validation'] == "K-Fold":
-        status, msg, Metrics = Validation.K_Fold(data)
+        # K-Fold
+        elif data['validation'] == "K-Fold":
+            status, msg, Metrics = MLA_Validation.K_Fold(data)
 
-    if status == "worked":
-        # Open json file for the experiment.
-        baseFolder = os.getcwd()
-        locationSavedResults = Path(baseFolder) / "SavedResults"
-        filename = secure_filename(data["projectName"] + ".json")
-        filepath = locationSavedResults / filename.lower()
-        if os.path.exists(filepath):
-            os.remove(filepath)
-            fp = open(filepath, "a")
-        else:
-            fp = open(filepath, "a")
+        if status == "worked":
+            # Open json file for the experiment.
+            baseFolder = os.getcwd()
+            locationSavedResults = Path(baseFolder) / "SavedResults"
+            filename = secure_filename(data["projectName"] + ".json")
+            filepath = locationSavedResults / filename.lower()
+            if os.path.exists(filepath):
+                os.remove(filepath)
+                fp = open(filepath, "a")
+            else:
+                fp = open(filepath, "a")
+    
+            # write to json file
+            metrics_Dump = json.dumps(Metrics)
 
-        # write to json file
-        metrics_Dump = json.dumps(Metrics)
+            fp.write(metrics_Dump)
 
-        fp.write(metrics_Dump)
+            # close the connection
+            fp.close()
 
-        # close the connection
-        fp.close()
+        Results = [status, msg, Metrics]
 
-    Results = [status, msg, Metrics]
+    # Perform Deep Learning Neural Network.
+    if data['methodology'] == "DLANN":
+        # Split
+        status, msg, Metrics = DLANN_Validation.Split(data)
+
+        if status == "worked":
+            # Open json file for the experiment.
+            baseFolder = os.getcwd()
+            locationSavedResults = Path(baseFolder) / "SavedResults"
+            filename = secure_filename(data["projectName"] + ".json")
+            filepath = locationSavedResults / filename.lower()
+            if os.path.exists(filepath):
+                os.remove(filepath)
+                fp = open(filepath, "a")
+            else:
+                fp = open(filepath, "a")
+    
+            # write to json file
+            metrics_Dump = json.dumps(Metrics)
+
+            fp.write(metrics_Dump)
+
+            # close the connection
+            fp.close()
+
+        Results = [status, msg, Metrics]
 
     return json.dumps(Results)
 
-
 @bp.route("/getResults", methods=["POST"])
-def getResults():
+def getResults(): 
     output = request.get_json()
     formData = json.loads(output)
 
@@ -284,7 +289,7 @@ def getResults():
 
     # close the connection
     fp.close()
-
+    
     return json.dumps(Metrics)
 
 
@@ -297,7 +302,6 @@ def getAlgorithmParameters():
 
     return json.dumps(Parameters)
 
-
 @bp.route("/getCategoryPreopts", methods=["POST"])
 def getCategoryPreopts():
     output = request.get_json()
@@ -306,6 +310,15 @@ def getCategoryPreopts():
     preopts = Preoptimization.getCategoryPreopts(data["Category"])
 
     return json.dumps(preopts)
+
+@bp.route("/getCategoryLayers", methods=["POST"])
+def getCategoryLayers():
+    output = request.get_json()
+    data = json.loads(output)
+
+    layers = NeuralNetwork.getCategoryLayers(data["Category"])
+
+    return json.dumps(layers)
 
 
 @bp.route("/results")
@@ -330,6 +343,16 @@ def getCSVResults():
 
     return json.dumps(Results)
 
+@bp.route("/getModelSummary", methods=["POST"])
+def getModelSummary():
+    processes_input = request.form.get("processes")
+    data = json.loads(processes_input)
+
+    status, msg, info = NeuralNetwork.getModelSummary(data)
+
+    Results = [status, msg, info]
+
+    return json.dumps(Results)
 
 @bp.route("/downloadCSV", methods=["POST"])
 def downloadCSV():
@@ -348,7 +371,6 @@ def downloadCSV():
 
     return json.dumps(Results)
 
-
 @bp.route("/getPreoptParameters", methods=["POST"])
 def getPreoptParameters():
     output = request.get_json()
@@ -357,7 +379,6 @@ def getPreoptParameters():
     Parameters = Preoptimization.getParameters(data["Preopt"])
 
     return json.dumps(Parameters)
-
 
 @bp.route("/getLayerParameters", methods=["POST"])
 def getLayerParameters():
@@ -369,8 +390,6 @@ def getLayerParameters():
     return json.dumps(Parameters)
 
 # Gets the colummn names inside the csv.
-
-
 @bp.route("/getCSVColumnTitles", methods=["POST"])
 def getCSVColumnTitles():
     filename = request.form.get("csvFileName")
@@ -381,3 +400,9 @@ def getCSVColumnTitles():
     columnTitles = Preoptimization.getCSVColumnTitles(data)
 
     return json.dumps(columnTitles)
+
+@bp.route("/getCompilerOptions", methods=["POST"])
+def getCompilerOptions():
+    compiler_Options = Compiler.getCompilerOptions()
+
+    return json.dumps(compiler_Options)
